move creditsSlice to features
