+++ src/store.js
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import movieSlice from "./features/Movies/movieSlice";
 import peopleSlice from "./features/People/peopleSlice";
 import genresSlice from "./features/Movies/genresSlice";
-import creditsSlice from "./features/Movies/MoviePage/creditsSlice";
+import creditsSlice from "./features/creditsSlice";
 import createSagaMiddleware from "redux-saga";
 import saga from "./saga";
